+++ src/headless/plugins/headlines/index.js
@@ -1,0 +1,62 @@
+/**
+ * @module converse-headlines
+ * @copyright 2022, the Converse.js contributors
+ * @description XEP-0045 Multi-User Chat Views
+ */
+import HeadlinesFeed from './feed.js';
+import headlines_api from './api.js';
+import { _converse, api, converse } from "@converse/headless/core";
+import { onHeadlineMessage } from './utils.js';
+
+
+converse.plugins.add('converse-headlines', {
+    /* Plugin dependencies are other plugins which might be
+     * overridden or relied upon, and therefore need to be loaded before
+     * this plugin.
+     *
+     * If the setting "strict_plugin_dependencies" is set to true,
+     * an error will be raised if the plugin is not found. By default it's
+     * false, which means these plugins are only loaded opportunistically.
+     *
+     * NB: These plugins need to have already been loaded via require.js.
+     */
+    dependencies: ["converse-chat"],
+
+    overrides: {
+        // Overrides mentioned here will be picked up by converse.js's
+        // plugin architecture they will replace existing methods on the
+        // relevant objects or classes.
+        //
+        // New functions which don't exist yet can also be added.
+
+        ChatBoxes: {
+            model (attrs, options) {
+                const { _converse } = this.__super__;
+                if (attrs.type == _converse.HEADLINES_TYPE) {
+                    return new _converse.HeadlinesFeed(attrs, options);
+                } else {
+                    return this.__super__.model.apply(this, arguments);
+                }
+            },
+        }
+    },
+
+
+    initialize () {
+        /**
+         * Shows headline messages
+         * @class
+         * @namespace _converse.HeadlinesFeed
+         * @memberOf _converse
+         */
+        _converse.HeadlinesFeed = HeadlinesFeed;
+
+        function registerHeadlineHandler () {
+            _converse.connection.addHandler(m => (onHeadlineMessage(m) || true), null, 'message');
+        }
+        api.listen.on('connected', registerHeadlineHandler);
+        api.listen.on('reconnected', registerHeadlineHandler);
+
+        Object.assign(api, headlines_api);
+    }
+});
