+++ src/shared/styles/themes/dracula.scss
@@ -23,6 +23,13 @@
 
     // ---
 
+    --headlines-color: var(--pink);
+    --headline-head-text-color: var(--headlines-color);
+    --headline-head-bg-color: var(--background);
+    --headline-message-color: var(--headlines-color);
+    --headline-separator-border-bottom: 2px solid var(--headlines-color);
+    --headline-head-border-bottom: 0.15em solid var(--headlines-color);
+
     --icon-hover-color: var(--cyan);
     --gray-color: var(--comment);
